check for presence of nodejs.toml before reading for clearing cache (#53)
* check for presence of nodejs.toml before reading for clearing cache

* change Node version to v16

* fix flag for detecting file and add log line

+++ buildpacks/nodejs/lib/build.sh
@@ -75,13 +75,19 @@ install_or_reuse_toolbox() {
 store_node_version() {
 	local layer_dir=$1
 	local prev_node_version
-	# shellcheck disable=SC2002
-	prev_node_version=$(cat "${layer_dir}.toml" | grep version | xargs | cut -d " " -f3)
-	mkdir -p "${layer_dir}/env.build"
-	if [[ -s "${layer_dir}/env.build/PREV_NODE_VERSION.override" ]]; then
-		rm -rf "${layer_dir}/env.build/PREV_NODE_VERSION.override"
+
+	if [[ -f "${layer_dir}.toml" ]]; then
+		# shellcheck disable=SC2002
+		prev_node_version=$(cat "${layer_dir}.toml" | grep version | xargs | cut -d " " -f3)
+		mkdir -p "${layer_dir}/env.build"
+
+		if [[ -s "${layer_dir}/env.build/PREV_NODE_VERSION.override" ]]; then
+			rm -rf "${layer_dir}/env.build/PREV_NODE_VERSION.override"
+		fi
+
+		info "Storing previous Node v${prev_node_version}"
+		echo -e "$prev_node_version\c" >"${layer_dir}/env.build/PREV_NODE_VERSION.override"
 	fi
-	echo -e "$prev_node_version\c" >>"${layer_dir}/env.build/PREV_NODE_VERSION.override"
 }
 
 install_or_reuse_node() {
@@ -95,7 +101,7 @@ install_or_reuse_node() {
 	status "Installing Node"
 	info "Getting Node version"
 	engine_node=$(json_get_key "$build_dir/package.json" ".engines.node")
-	node_version=${engine_node:-14.x}
+	node_version=${engine_node:-16.x}
 
 	info "Resolving Node version"
 	resolved_data=$(resolve-version node "$node_version")
@@ -110,9 +116,8 @@ install_or_reuse_node() {
 		mkdir -p "${layer_dir}"
 		rm -rf "${layer_dir:?}"/*
 
-		echo "cache = true" >"${layer_dir}.toml"
-
 		{
+			echo "cache = true"
 			echo "build = true"
 			echo "launch = true"
 			echo -e "[metadata]\nversion = \"$node_version\""
